refactor(firestore): hoist shared routing state, types, and snapshot helpers
First step of splitting the pyric/firestore entry into per-family
modules: lift the cross-cutting machinery into three sibling modules
before any operation moves, so the family extractions that follow are
pure moves.

- state.ts: the TARGET_SYMBOL brand, ref/converter/underlying WeakMaps,
  tag/resolve routing helpers, chain/fb cast shims, and the sandbox
  read-path value finalizers. Package-internal except TARGET_SYMBOL.
- types.ts: the public handle, reference, query, snapshot, converter,
  and batch/transaction types plus the scalar/util type re-exports.
- snapshots.ts: sandbox snapshot rehydration shared by reads + listeners
  (value wrapping, ref tagging, exists property to method normalization).

The entry imports these and re-exports the public types + brand; every
inline operation still lives in index.ts. No behavior change: the frozen
runtime export surface and the full typed export surface are identical.
